Move ParallelTest class to testing.
