+++ tests/test_parallel.py
@@ -13,52 +13,14 @@ tests for parallel iterators
 # The full license is in the file COPYING.txt, distributed with this software.
 #-----------------------------------------------------------------------------
 
-from numpy.testing import assert_array_equal
 import os
-import sys
-import unittest
-import ytree
 
-from ytree.utilities.testing import TempDirTest
+from ytree.utilities.testing import ParallelTest, TempDirTest
 
-try:
-    from mpi4py import MPI
-except ModuleNotFoundError:
-    MPI = None
-
-class ParallelTest:
-    base_filename = "tiny_ctrees/locations.dat"
-    test_filename = "test_arbor/test_arbor.h5"
-    test_script = None
-    ncores = 4
-
-    @property
-    def test_script_path(self):
-        return os.path.join(os.path.dirname(__file__), self.test_script)
-
-    def check_values(self, arbor, my_args):
-        group = my_args[0]
-        assert_array_equal(arbor["test_field"], 2 * arbor["mass"])
-        for tree in arbor:
-            assert_array_equal(tree[group, "test_field"], 2 * tree[group, "mass"])
-
-    @unittest.skipIf(MPI is None, "mpi4py not installed")
-    def test_parallel(self):
-
-        for i, my_args in enumerate(self.arg_sets):
-            with self.subTest(i=i):
-
-                # test_data_path = self.prepare_data()
-                args = [self.test_script_path, self.base_filename, self.test_filename] + \
-                    [str(arg) for arg in my_args]
-                comm = MPI.COMM_SELF.Spawn(sys.executable, args=args, maxprocs=self.ncores)
-                comm.Disconnect()
-
-                test_arbor = ytree.load(self.test_filename)
-                self.check_values(test_arbor, my_args)
+script_path = os.path.dirname(__file__)
 
 class ParallelTreesTest(TempDirTest, ParallelTest):
-    test_script = "parallel/parallel_trees.py"
+    test_script = os.path.join(script_path, "parallel/parallel_trees.py")
     arg_sets = (
         ("forest", 0, 0, 4),
         ("tree",   0, 0, 4),
@@ -69,7 +31,7 @@ class ParallelTreesTest(TempDirTest, ParallelTest):
     )
 
 class ParallelTreeNodesTest(TempDirTest, ParallelTest):
-    test_script = "parallel/parallel_tree_nodes.py"
+    test_script = os.path.join(script_path, "parallel/parallel_tree_nodes.py")
     arg_sets = (
         ("forest", 0, 0),
         ("tree",   0, 0),
@@ -79,7 +41,7 @@ class ParallelTreeNodesTest(TempDirTest, ParallelTest):
     )
 
 class ParallelNodesTest(TempDirTest, ParallelTest):
-    test_script = "parallel/parallel_nodes.py"
+    test_script = os.path.join(script_path, "parallel/parallel_nodes.py")
     arg_sets = (
         ("forest", 1, 0, 0, 0, 4),
         ("tree",   1, 0, 0, 0, 4),
